\name{compareGenotypes}
\alias{compareGenotypes}
\alias{compareGenotypes.default}
\title{compareGenotypes}
\description{compareGenotypes}
\usage{
compareGenotypes(crossData, outputPath, cutoffP)
}
\arguments{
  \item{crossData}{a}
  \item{outputPath}{a}
  \item{cutoffP}{a}
}
\author{Rose Imee Zhella Morantte}
\keyword{analysis}
