% Generated by roxygen2 (4.0.2): do not edit by hand
\name{connectAccess64}
\alias{connectAccess64}
\title{Connect to Access DB using RODBC, for 64bit windows}
\usage{
connectAccess64(dbpath)
}
\arguments{
\item{dbpath}{path to database}
}
\value{
connection
}
\description{
Connect to Access DB using RODBC, for 64bit windows
}
\examples{
connectAccess('A:/database.accdb')
}
\keyword{64bit}
\keyword{Access,}
\keyword{RODBC,}
\keyword{database,}
\keyword{odbc,}

