\name{createQTLdata}
\alias{createQTLdata}
\alias{createQTLdata.default}
\title{createQTLdata}
\description{createQTLdata}
\usage{
createQTLdata(outputPath, dataFormat, format1 = NULL, crossType = "f2", file1, format2 = NULL, file2 = NULL, format3 = NULL, file3 = NULL, P_geno = NULL, bcNum = 0, fNum = 0)
}
\arguments{
  \item{outputPath}{a}
  \item{dataFormat}{a}
  \item{format1}{a}
  \item{crossType}{a}
  \item{file1}{a}
  \item{format2}{a}
  \item{file2}{a}
  \item{format3}{a}
  \item{file3}{a}
  \item{P_geno}{a}
  \item{bcNum}{a}
  \item{fNum}{a}
}
\author{Rose Imee Zhella Morantte}
\keyword{analysis}