\name{designLattice}
\alias{designLattice}
\alias{designLattice.default}
\title{Randomization for Lattice Design}
\description{description1}
\usage{
designLattice(generate, r = 2, trial = 1, numFieldRow = 1, serpentine = FALSE, file = NULL)
}
\arguments{
  \item{generate}{a list}
  \item{r}{numeric}
  \item{trial}{numeric}
  \item{numFieldRow}{numeric}
  \item{serpentine}{logical}
  \item{file}{NULL}
}
\author{Alaine A. Gulles}
\keyword{ design }
