\name{Diallel_M4}
\alias{Diallel_M4}
\docType{data}
\title{Diallel_M4}
\description{
data for Diallel 4
}
\usage{data(Diallel_M4)}
\format{
  A data frame with 126 observations on the following 7 variables.
  \describe{
    \item{\code{Env}}{a factor with levels \code{Normal} \code{Saline}}
    \item{\code{P1}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6}}
    \item{\code{P2}}{a factor with levels \code{2} \code{3} \code{4} \code{5} \code{6} \code{7}}
    \item{\code{Block}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{Plant_height}}{a numeric vector}
    \item{\code{Grain_yield}}{a numeric vector}
    \item{\code{Th_grain_weight}}{a numeric vector}
  }
}
\examples{
data(Diallel_M4)
## maybe str(Diallel_M4) ; plot(Diallel_M4) ...
}
\keyword{datasets}
