\name{doIM}
\alias{doIM}
\alias{doIM.default}
\title{doIM}
\description{doIM}
\usage{
doIM(outputPath, crossData2, yVars, stepCalc = 0, errCalc = 0.01, mapCalc = c("haldane", "kosambi", "c-f", "morgan"), lodCutoffM = 3, phenoModel = c("normal", "binary", "2part", "np"), alMethod = c("em", "imp", "hk", "ehk"), nPermutations = 100)
}
\arguments{
  \item{outputPath}{a}
  \item{crossData2}{a}
  \item{yVars}{a}
  \item{stepCalc}{a}
  \item{errCalc}{a}
  \item{mapCalc}{a}
  \item{lodCutoffM}{a}
  \item{phenoModel}{a}
  \item{alMethod}{a}
  \item{nPermutations}{a}
}
\author{Rose Imee Zhella Morantte}
\keyword{analysis}
