\name{doQTL_CIM}
\alias{doQTL_CIM}
\alias{doQTL_CIM.default}
\title{doQTL_CIM}
\description{doQTL_CIM}
\usage{
doQTL_CIM(outputPath, crossData2, yVars, stepCalc = 0, errCalc = 0.01, mapCalc = c("haldane", "kosambi", "c-f", "morgan"), lodCutoffM = 3, alMethod = c("em", "imp", "hk", "ehk"), nPermutations = 100, numCovar = 1, winSize = 10)
}
\arguments{
  \item{outputPath}{a}
  \item{crossData2}{a}
  \item{yVars}{a}
  \item{stepCalc}{a}
  \item{errCalc}{a}
  \item{mapCalc}{a}
  \item{lodCutoffM}{a}
  \item{alMethod}{a}
  \item{nPermutations}{a}
  \item{numCovar}{a}
  \item{winSize}{a}
}
\author{Rose Imee Zhella Morantte}
\keyword{analysis}