\name{doQTL_MQM}
\alias{doQTL_MQM}
\alias{doQTL_MQM.default}
\title{doQTL_MQM}
\description{doQTL_MQM}
\usage{
doQTL_MQM(outputPath, crossData2, yVars, lodCutoffM = 3, nPermutations = 100, winSize = 10, stepSize = 5, addModel = TRUE, numCofac = 1, mlAlgo = TRUE)
}
\arguments{
  \item{outputPath}{a}
  \item{crossData2}{a}
  \item{yVars}{a}
  \item{lodCutoffM}{a}
  \item{nPermutations}{a}
  \item{winSize}{a}
  \item{stepSize}{a}
  \item{addModel}{a}
  \item{numCofac}{q}
  \item{mlAlgo}{a}
}
\author{Rose Imee Zhella Morantte}
\keyword{analysis}