\name{fac.gen}
\alias{fac.gen}
\title{fac.gen}
\description{
 generates complete list of treatment combinations for factorial design. This function is obtained from the dae package which is authoured by Chris Brien
}
\usage{
fac.gen(generate, each = 1, times = 1, order = "standard")
}
\arguments{
  \item{generate}{a}
  \item{each}{a}
  \item{times}{a}
  \item{order}{a}
}
\author{Chris Brien of dae package}
\keyword{factorial}

