\name{FinlayStability}
\alias{FinlayStability}
\alias{FinlayStability.default}
\title{Finlay-Wilkinson Stability Analysis}
\description{Stability Analysis using Finlay-Wilkinson Method. Computation is based on Statistical and
Biometrical Techniques in Plant Breeding by Jawahar R. Sharma}
\usage{
FinlayStability(data, respvar, geno, env, outputPath = NULL)
}
\arguments{
  \item{data}{dataframe}
  \item{respvar}{character string denoting the response variable}
  \item{geno}{character string denoting the genotypic variable}
  \item{env}{character string denoting the environment variable}
  \item{outputPath}{path where graphical output will be save}
}
\author{VIBartolome, AAGulles}
\keyword{stability}
\keyword{finlay-wilkinson}

