\name{formatAnovaTable}
\alias{formatAnovaTable}
\alias{formatAnovaTable.default}
\title{
format Anova Table
}
\description{
formats a number to a specific format
}
\usage{
formatAnovaTable(anova.table)
}
\arguments{
  \item{anova.table}{the table that will be formatted }
}
\author{
Nellwyn Sales
}
\keyword{formatDecimal}
