\name{formatNumericValue}
\alias{formatNumericValue}
\alias{formatNumericValue.default}
\title{
format Numeric Value
}
\description{
formats a number to a specific format
}
\usage{
formatNumericValue(number)
}
\arguments{
  \item{number}{the number that will be formatted }
}
\author{
Nellwyn Sales
}
\keyword{formatDecimal}
