\name{genDistance}
\alias{genDistance}
\alias{genDistance.default}
\title{
genDistance
}
\description{
genDistance
}
\usage{
genDistance(genindObject, method=1, displayDiag=FALSE, displayUpper=FALSE)
}
\arguments{
  \item{genindObject}{genind object}
  \item{method}{integer that corresponds to the method to be used}
  \item{displayDiag}{logical}
  \item{displayUpper}{logical}
}
\author{
Nellwyn Sales
}
\keyword{geneticDistance}
