\name{GenMean_raw}
\alias{GenMean_raw}
\docType{data}
\title{GenMean_raw}
\description{
raw data for generation mean analysis
}
\usage{data(GenMean_raw)}
\format{
  A data frame with 15 observations on the following 8 variables.
  \describe{
    \item{\code{rep}}{a numeric vector}
    \item{\code{P1}}{a numeric vector}
    \item{\code{P2}}{a numeric vector}
    \item{\code{F1}}{a numeric vector}
    \item{\code{F2}}{a numeric vector}
    \item{\code{F3}}{a numeric vector}
    \item{\code{BC1}}{a numeric vector}
    \item{\code{BC2}}{a numeric vector}
  }
}
\examples{
data(GenMean_raw)
## maybe str(GenMean_raw) ; plot(GenMean_raw) ...
}
\keyword{datasets}
