\name{GenMean_summaryStats}
\alias{GenMean_summaryStats}
\docType{data}
\title{GenMean_summaryStats}
\description{
data on summary statistics for generation mean analysis
}
\usage{data(GenMean_summaryStats)}
\format{
  A data frame with 6 observations on the following 5 variables.
  \describe{
    \item{\code{generation}}{a factor with levels \code{BC1} \code{BC2} \code{F1} \code{F2} \code{P1} \code{P2}}
    \item{\code{mean}}{a numeric vector}
    \item{\code{stddev}}{a numeric vector}
    \item{\code{n}}{a numeric vector}
    \item{\code{Weights}}{a numeric vector}
  }
}
\examples{
data(GenMean_summaryStats)
## maybe str(GenMean_summaryStats) ; plot(GenMean_summaryStats) ...
}
\keyword{datasets}
