\name{GETwoStage_resid}
\alias{GETwoStage_resid}
\alias{GETwoStage_resid.default}
\title{GETwoStage_resid}
\description{
This function consolidates the residuals generated by running GETwoStage.test
}
\usage{
GETwoStage_resid(ge2sout, respvar, is.genoRandom = FALSE)
}
\arguments{
  \item{ge2sout}{object generated after running GETwoStage.test}
  \item{respvar}{a vector of strings; variable names of the response variables}
  \item{is.genoRandom}{logical; indicating whether genotype/treatment is random or not; default value is FALSE (FIXED factor)}
}
\author{
Alaine A. Gulles, Rose Imee Zhella Morantte, Nellwyn L. Sales
}
\keyword{residual}
