\name{graph.mea1s.predint}
\alias{graph.mea1s.predint}
\alias{graph.mea1s.predint.default}
\title{graph.mea1s.predint}
\description{
prediction interval
}
\usage{
graph.mea1s.predint(data, respVar, geno, result)
}
\arguments{
  \item{data}{a}
  \item{respVar}{a}
  \item{geno}{a}
  \item{result}{a}
}
\author{
Alaine A. Gulles
}
\keyword{graph}
