\name{graph.mea2s.boxhist}
\alias{graph.mea2s.boxhist}
\alias{graph.mea2s.boxhist.default}
\title{graph.mea2s.boxhist}
\description{
boxhist
}
\usage{
graph.mea2s.boxhist(data, respvar, result, box = FALSE, hist = FALSE)
}
\arguments{
  \item{data}{a}
  \item{respvar}{a}
  \item{result}{a}
  \item{box}{a}
  \item{hist}{a}
}
\author{
Alaine A. Gulles
}
\keyword{ ~kwd1 }
