\name{graph.mea2s.diagplots}
\alias{graph.mea2s.diagplots}
\alias{graph.mea2s.diagplots.default}
\title{graph.mea2s.diagplots}
\description{
diagnostic plots
}
\usage{
graph.mea2s.diagplots(data, respvar, is.random = FALSE, result)
}
\arguments{
  \item{data}{a}
  \item{respvar}{a}
  \item{is.random}{a}
  \item{result}{a}
}
\author{
Alaine A. Gulles
}
\keyword{graph}
