\name{graph.mea2s.predint}
\alias{graph.mea2s.predint}
\alias{graph.mea2s.predint.default}
\title{graph.mea2s.predint}
\description{
prediction interval
}
\usage{
graph.mea2s.predint(data, respvar, geno, result)
}
\arguments{
  \item{data}{a}
  \item{respvar}{a}
  \item{geno}{a}
  \item{result}{a}
}
\author{
Alaine A. Gulles
}
\keyword{graph}