% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{ifgbasecolours}
\alias{ifgbasecolours}
\title{Colour set for theme - IfG corporate colours - 7 base colours only}
\format{Matrix with 1 column and 7 rows - 7 base colours only}
\usage{
data(themecolours)
}
\description{
Colour set for theme - IfG corporate colours - 7 base colours only
}
\keyword{datasets}

