% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ifgcolours}
\alias{ifgcolours}
\title{Colour set for theme - IfG corporate colours - 4 tints of 7 colours}
\format{Matrix with 4 columns and 7 rows - 4 tints (25\% steps) of 7 colours}
\usage{
data(themecolours)
}
\description{
Colour set for theme - IfG corporate colours - 4 tints of 7 colours
}
\keyword{datasets}

