% Generated by roxygen2 (4.0.1): do not edit by hand
\name{knit2byrokrates}
\alias{knit2byrokrates}
\title{Rmd to Byrokrates}
\usage{
knit2byrokrates(filename, title = paste("R blog", format(Sys.time(),
  "\%d/\%m/\%Y \%H:\%M")), publish = F, preview = T)
}
\arguments{
\item{filename}{name od Rmd file}

\item{title}{Title of blog post}

\item{publish}{FALSE for draft, TRUE to publish. Defaults to FALSE}

\item{preview}{Whether to open preview in browser upon publication. Defaults to TRUE}
}
\description{
Upload Rmd document as post to Byrokrates WordPress blog
}
\keyword{blog}
\keyword{wordpress}

