\name{load.cross.data}
\alias{load.cross.data}
\title{load.cross.data}
\description{
a
}
\usage{
load.cross.data(P.data, G.data, map.data, cross, heterozygotes = TRUE, genotype, env.label = NULL, env = NULL, ST = NULL)
}
\arguments{
  \item{P.data}{a}
  \item{G.data}{a}
  \item{map.data}{a}
  \item{cross}{a}
  \item{heterozygotes}{a}
  \item{genotype}{a}
  \item{env.label}{a}
  \item{env}{a}
  \item{ST}{a}
}
\author{
Paul Eilers, Gerrit Gort, Sabine Schnabel, Lucia Gutierrez, Marcos Malosetti, Joost van Heerwaarden, Fred van Eeuwijk
}
\keyword{cross}
