% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{LoadCustomThemes}
\alias{LoadCustomThemes}
\title{Load custom theme}
\usage{
LoadCustomThemes(mycols, fontfamily = "Helvetica")
}
\arguments{
\item{mycols}{A vector of colours in hex (HTML) format.}

\item{fontfamily}{Font family. Defaults to 'Helvetica'}
}
\description{
Create custom ggplot theme based on a set of colours and a font family
}
\examples{
LoadCustomThemes()
}
\keyword{ggplot,}
\keyword{themes}

