\name{MESIMIndex_2}
\alias{MESIMIndex_2}
\title{MESIMIndex_2}
\description{
a
}
\usage{
MESIMIndex_2(file.dat = NULL, weights = NULL, selval = 5, design = "lattice", corr = FALSE, rawdata = TRUE, markers, qtls)
}
\arguments{
  \item{file.dat}{a}
  \item{weights}{a}
  \item{selval}{a}
  \item{design}{a}
  \item{corr}{a}
  \item{rawdata}{a}
  \item{markers}{a}
  \item{qtls}{a}
}
\author{
S. Perez-Elizalde, J. Crossa, J. Ceron-Rojas, and G. Alvarado
}
\keyword{a}
