\name{MQ.marker.diag}
\alias{MQ.marker.diag}
\title{MQ.marker.diag}
\description{
for QTL analysis
}
\usage{
MQ.marker.diag(crossobj, QTL.path, estmarker = FALSE, thresholdMQ = 0.1, quantile = FALSE, p.val = 0.01, na.cutoff = 0.1)
}
\arguments{
  \item{crossobj}{a}
  \item{QTL.path}{a}
  \item{estmarker}{a}
  \item{thresholdMQ}{a}
  \item{quantile}{a}
  \item{p.val}{a}
  \item{na.cutoff}{a}
}
\author{Paul Eilers, Gerrit Gort, Sabine Schnabel, Lucia Gutierrez, Marcos Malosetti, Joost van Heerwaarden, Fred van Eeuwijk}
\keyword{marker}
