\name{NCI_ME}
\alias{NCI_ME}
\docType{data}
\title{NCI_ME}
\description{
data for North Carolina Experiment I
}
\usage{data(NCI_ME)}
\format{
  A data frame with 176 observations on the following 5 variables.
  \describe{
    \item{\code{Env}}{a factor with levels \code{A} \code{B}}
    \item{\code{Male}}{a factor with levels \code{1} \code{2} \code{3} \code{4}}
    \item{\code{Female}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9} \code{10} \code{11}}
    \item{\code{Block}}{a factor with levels \code{1} \code{2}}
    \item{\code{Y}}{a numeric vector}
  }
}
\examples{
data(NCI_ME)
## maybe str(NCI_ME) ; plot(NCI_ME) ...
}
\keyword{datasets}
