\name{NCII_ME}
\alias{NCII_ME}
\docType{data}
\title{NCII_ME}
\description{
data for North Carolina Experiment II
}
\usage{data(NCII_ME)}
\format{
  A data frame with 384 observations on the following 5 variables.
  \describe{
    \item{\code{Env}}{a factor with levels \code{A} \code{B}}
    \item{\code{Female}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8}}
    \item{\code{Male}}{a factor with levels \code{9} \code{10} \code{11} \code{12} \code{13} \code{14} \code{15} \code{16}}
    \item{\code{Block}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{Y}}{a numeric vector}
  }
}
\examples{
data(NCII_ME)
## maybe str(NCII_ME) ; plot(NCII_ME) ...
}
\keyword{datasets}
