\name{NCIII_ME}
\alias{NCIII_ME}
\docType{data}
\title{NCIII_ME}
\description{
data for North Carolina Experiment III
}
\usage{data(NCIII_ME)}
\format{
  A data frame with 128 observations on the following 5 variables.
  \describe{
    \item{\code{Env}}{a factor with levels \code{A} \code{B}}
    \item{\code{Male}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9} \code{10} \code{11} \code{12} \code{13} \code{14} \code{15} \code{16}}
    \item{\code{Female}}{a factor with levels \code{1} \code{2}}
    \item{\code{Block}}{a factor with levels \code{1} \code{2}}
    \item{\code{Y}}{a numeric vector}
  }
}
\examples{
data(NCIII_ME)
## maybe str(NCIII_ME) ; plot(NCIII_ME) ...
}
\keyword{datasets}
