\name{PBTools-package}
\alias{PBTools-package}
\alias{PBTools}
\docType{package}
\title{
Plant Breeding Tools
}
\description{
Plant Breeding Tools for single and multiple environment analysis. And on 
Version 1.1 is surpporting Introgression Line, Pyramided Line and SSSL.
}
\details{
\tabular{ll}{
Package: \tab PBTools\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-01-06\cr
License: \tab GPL-2\cr
}
}
\author{
BBI Group of IRRI

Maintainer: MAO QIN <MQIN@YMAIL.COM>
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
