\name{popStructure}
\alias{popStructure}
\alias{popStructure.default}
\title{
 This function computes fst, fit, fis and pairwise fst. This uses the fstat function of adegenet package.
}
\description{
popStructure
}
\usage{
popStructure(genindObject, pairwiseFst=FALSE)
}
\arguments{
  \item{genindObject}{genind object}
  \item{pairwiseFst}{logical}
}
\author{
Nellwyn Sales
}
\keyword{popStructure}
