\name{QTL.analysis}
\alias{QTL.analysis}
\title{QTL.analysis}
\description{
performs QTL analysis
}
\usage{
QTL.analysis(crossobj = cross.data, QTL.path, env.label = NULL, env = NULL, trait = "pred", step, method, threshold, distance, cofactors, window.size)
}
\arguments{
  \item{crossobj}{a}
  \item{QTL.path}{a}
  \item{env.label}{a}
  \item{env}{a}
  \item{trait}{a}
  \item{step}{a}
  \item{method}{a}
  \item{threshold}{a}
  \item{distance}{a}
  \item{cofactors}{a}
  \item{window.size}{a}
}
\author{Paul Eilers, Gerrit Gort, Sabine Schnabel, Lucia Gutierrez, Marcos Malosetti, Joost van Heerwaarden, Fred van Eeuwijk}
\keyword{qtl}
