\name{QTL_pred_height}
\alias{QTL_pred_height}
\docType{data}
\title{QTL_pred_height}
\description{
data on predicted values of traits for QTL analysis
}
\usage{data(QTL_pred_height)}
\format{
  A data frame with 202 observations on the following 4 variables.
  \describe{
    \item{\code{genotype}}{a factor with levels \code{G_001} \code{G_002} \code{G_003} \code{G_004} \code{G_005} \code{G_006} \code{G_007} \code{G_008} \code{G_009} \code{G_010} \code{G_011} \code{G_012} \code{G_013} \code{G_014} \code{G_015} \code{G_016} \code{G_017} \code{G_018} \code{G_019} \code{G_020} \code{G_021} \code{G_022} \code{G_023} \code{G_024} \code{G_025} \code{G_026} \code{G_027} \code{G_028} \code{G_029} \code{G_030} \code{G_031} \code{G_032} \code{G_033} \code{G_034} \code{G_035} \code{G_036} \code{G_037} \code{G_038} \code{G_039} \code{G_040} \code{G_041} \code{G_042} \code{G_043} \code{G_044} \code{G_045} \code{G_046} \code{G_047} \code{G_048} \code{G_049} \code{G_050} \code{G_051} \code{G_052} \code{G_053} \code{G_054} \code{G_055} \code{G_056} \code{G_057} \code{G_058} \code{G_059} \code{G_060} \code{G_061} \code{G_062} \code{G_063} \code{G_064} \code{G_065} \code{G_066} \code{G_067} \code{G_068} \code{G_069} \code{G_070} \code{G_071} \code{G_072} \code{G_073} \code{G_074} \code{G_075} \code{G_076} \code{G_077} \code{G_078} \code{G_079} \code{G_080} \code{G_081} \code{G_082} \code{G_083} \code{G_084} \code{G_085} \code{G_086} \code{G_087} \code{G_088} \code{G_089} \code{G_090} \code{G_091} \code{G_092} \code{G_093} \code{G_094} \code{G_095} \code{G_096} \code{G_097} \code{G_098} \code{G_099} \code{G_100} \code{G_101} \code{G_102} \code{G_103} \code{G_104} \code{G_105} \code{G_106} \code{G_107} \code{G_108} \code{G_109} \code{G_110} \code{G_111} \code{G_112} \code{G_113} \code{G_114} \code{G_115} \code{G_116} \code{G_117} \code{G_118} \code{G_119} \code{G_120} \code{G_121} \code{G_122} \code{G_123} \code{G_124} \code{G_125} \code{G_126} \code{G_127} \code{G_128} \code{G_129} \code{G_130} \code{G_131} \code{G_132} \code{G_133} \code{G_134} \code{G_135} \code{G_136} \code{G_137} \code{G_138} \code{G_139} \code{G_140} \code{G_141} \code{G_142} \code{G_143} \code{G_144} \code{G_145} \code{G_146} \code{G_147} \code{G_148} \code{G_149} \code{G_150} \code{G_151} \code{G_152} \code{G_153} \code{G_154} \code{G_155} \code{G_156} \code{G_157} \code{G_158} \code{G_159} \code{G_160} \code{G_161} \code{G_162} \code{G_163} \code{G_164} \code{G_165} \code{G_166} \code{G_167} \code{G_168} \code{G_169} \code{G_170} \code{G_171} \code{G_172} \code{G_173} \code{G_174} \code{G_175} \code{G_176} \code{G_177} \code{G_178} \code{G_179} \code{G_180} \code{G_181} \code{G_182} \code{G_183} \code{G_184} \code{G_185} \code{G_186} \code{G_187} \code{G_188} \code{G_189} \code{G_190} \code{G_191} \code{G_192} \code{G_193} \code{G_194} \code{G_195} \code{G_196} \code{G_197} \code{G_198} \code{G_199} \code{G_200} \code{P1} \code{P2}}
    \item{\code{pred}}{a numeric vector}
    \item{\code{ue}}{a numeric vector}
    \item{\code{env}}{a factor with levels \code{1}}
  }
}
\examples{
data(QTL_pred_height)
## maybe str(QTL_pred_height) ; plot(QTL_pred_height) ...
}
\keyword{datasets}
