\name{QTLDataPrep}
\alias{QTLDataPrep}
\alias{QTLDataPrep.default}
\title{QTLDataPrep}
\description{QTLDataPrep}
\usage{
QTLDataPrep(P_data, G_data, M_data, P_geno)
}
\arguments{
  \item{P_data}{a}
  \item{G_data}{a}
  \item{M_data}{a}
  \item{P_geno}{a}
}
\author{Rose Imee Zhella Morantte}
\keyword{data}

