% Generated by roxygen2 (4.0.2): do not edit by hand
\name{renamevars}
\alias{renamevars}
\title{Rename multiple variables in a df}
\usage{
renamevars(dataset, renamelist, renamelistfile)
}
\arguments{
\item{dataset}{Data set whose column names are to be changed}

\item{renamelist}{List of two vectors of same lenght, or df with two columns}

\item{csv}{with two columns}
}
\description{
Rename multiple df variables based on two lists of values. If a name contained
in names(dataset) is not found in renamelist[1], it is left as is.
}
\examples{
renamevars(dataframe, renamelist = list(c('oldname'),c('newname'))
renamevars(dataframe, renamelistfile = './tableofnames.csv')
}
\keyword{data.frame,}
\keyword{names}
\keyword{rename,}
\keyword{variables,}

