% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rgb2col}
\alias{rgb2col}
\title{Load custom theme}
\usage{
rgb2col(rgb)
}
\arguments{
\item{rgb}{Vector with \% RGB values}
}
\value{
string
}
\description{
Create hex value of colour based on RGB value
}
\examples{
rgb(1,1,1) # returns #FFFFFF
}
\keyword{ggplot,}
\keyword{themes}

