% Generated by roxygen2 (4.0.2): do not edit by hand
\name{savefilewithdate}
\alias{savefilewithdate}
\title{Save file with timestamp, with option for time-stamped}
\usage{
savefilewithdate()
}
\arguments{
\item{object}{Ggplot object.}

\item{plotname}{String. Defaults 'Plot'.}

\item{ffamily}{Font family. Defaults to 'Helvetica'}

\item{plotformat}{String}

\item{plotw}{Width of plot in centimeters. Defaults to 14}

\item{ploth}{Height of plot in centimeters. Defaults to 11.5}

\item{plotdir}{Directory for saving - plot, data and image are saved into subdirectories}
}
\description{
Save object into file
}
\examples{
savefilewithdate()
}
\keyword{data,}
\keyword{date}
\keyword{file,}
\keyword{save,}

