% Generated by roxygen2 (4.0.2): do not edit by hand
\name{saveplot}
\alias{saveplot}
\title{Save plot with ggsave}
\usage{
saveplot(plotname = "Plot", plotformat = "eps", ffamily = "Helvetica",
  splot = last_plot(), ploth = 21/2, plotw = 14, plotdir = ".",
  dpi = 96)
}
\arguments{
\item{plotname}{String. Defaults 'Plot'.}

\item{plotformat}{String}

\item{ffamily}{Font family. Defaults to 'Helvetica'}

\item{splot}{Ggplot object. Defaults to last plot}

\item{ploth}{Height of plot in centimeters. Defaults to 11.5}

\item{plotw}{Width of plot in centimeters. Defaults to 14}

\item{plotdir}{Directory for saving - plot, data and image are saved into subdirectories}

\item{dpi}{resolution for raster graphics, passed to ggsave}
}
\description{
Save ggplot - ggsave wrapper. Saves image, GGplot object, and data.
}
\examples{
saveplot()
}
\keyword{ggplot,}
\keyword{themes}

