\name{SI_markers}
\alias{SI_markers}
\docType{data}
\title{SI_markers}
\description{
data on markers for Selection Index Analysis
}
\usage{data(SI_markers)}
\format{
  A data frame with 210 observations on the following 126 variables.
  \describe{
    \item{\code{V}}{a numeric vector}
    \item{\code{M1}}{a numeric vector}
    \item{\code{M2}}{a numeric vector}
    \item{\code{M3}}{a numeric vector}
    \item{\code{M4}}{a numeric vector}
    \item{\code{M5}}{a numeric vector}
    \item{\code{M6}}{a numeric vector}
    \item{\code{M7}}{a numeric vector}
    \item{\code{M8}}{a numeric vector}
    \item{\code{M9}}{a numeric vector}
    \item{\code{M10}}{a numeric vector}
    \item{\code{M11}}{a numeric vector}
    \item{\code{M12}}{a numeric vector}
    \item{\code{M13}}{a numeric vector}
    \item{\code{M14}}{a numeric vector}
    \item{\code{M15}}{a numeric vector}
    \item{\code{M16}}{a numeric vector}
    \item{\code{M17}}{a numeric vector}
    \item{\code{M18}}{a numeric vector}
    \item{\code{M19}}{a numeric vector}
    \item{\code{M20}}{a numeric vector}
    \item{\code{M21}}{a numeric vector}
    \item{\code{M22}}{a numeric vector}
    \item{\code{M23}}{a numeric vector}
    \item{\code{M24}}{a numeric vector}
    \item{\code{M25}}{a numeric vector}
    \item{\code{M26}}{a numeric vector}
    \item{\code{M27}}{a numeric vector}
    \item{\code{M28}}{a numeric vector}
    \item{\code{M29}}{a numeric vector}
    \item{\code{M30}}{a numeric vector}
    \item{\code{M31}}{a numeric vector}
    \item{\code{M32}}{a numeric vector}
    \item{\code{M33}}{a numeric vector}
    \item{\code{M34}}{a numeric vector}
    \item{\code{M35}}{a numeric vector}
    \item{\code{M36}}{a numeric vector}
    \item{\code{M37}}{a numeric vector}
    \item{\code{M38}}{a numeric vector}
    \item{\code{M39}}{a numeric vector}
    \item{\code{M40}}{a numeric vector}
    \item{\code{M41}}{a numeric vector}
    \item{\code{M42}}{a numeric vector}
    \item{\code{M43}}{a numeric vector}
    \item{\code{M44}}{a numeric vector}
    \item{\code{M45}}{a numeric vector}
    \item{\code{M46}}{a numeric vector}
    \item{\code{M47}}{a numeric vector}
    \item{\code{M48}}{a numeric vector}
    \item{\code{M49}}{a numeric vector}
    \item{\code{M50}}{a numeric vector}
    \item{\code{M51}}{a numeric vector}
    \item{\code{M52}}{a numeric vector}
    \item{\code{M53}}{a numeric vector}
    \item{\code{M54}}{a numeric vector}
    \item{\code{M55}}{a numeric vector}
    \item{\code{M56}}{a numeric vector}
    \item{\code{M57}}{a numeric vector}
    \item{\code{M58}}{a numeric vector}
    \item{\code{M59}}{a numeric vector}
    \item{\code{M60}}{a numeric vector}
    \item{\code{M61}}{a numeric vector}
    \item{\code{M62}}{a numeric vector}
    \item{\code{M63}}{a numeric vector}
    \item{\code{M64}}{a numeric vector}
    \item{\code{M65}}{a numeric vector}
    \item{\code{M66}}{a numeric vector}
    \item{\code{M67}}{a numeric vector}
    \item{\code{M68}}{a numeric vector}
    \item{\code{M69}}{a numeric vector}
    \item{\code{M70}}{a numeric vector}
    \item{\code{M71}}{a numeric vector}
    \item{\code{M72}}{a numeric vector}
    \item{\code{M73}}{a numeric vector}
    \item{\code{M74}}{a numeric vector}
    \item{\code{M75}}{a numeric vector}
    \item{\code{M76}}{a numeric vector}
    \item{\code{M77}}{a numeric vector}
    \item{\code{M78}}{a numeric vector}
    \item{\code{M79}}{a numeric vector}
    \item{\code{M80}}{a numeric vector}
    \item{\code{M81}}{a numeric vector}
    \item{\code{M82}}{a numeric vector}
    \item{\code{M83}}{a numeric vector}
    \item{\code{M84}}{a numeric vector}
    \item{\code{M85}}{a numeric vector}
    \item{\code{M86}}{a numeric vector}
    \item{\code{M87}}{a numeric vector}
    \item{\code{M88}}{a numeric vector}
    \item{\code{M89}}{a numeric vector}
    \item{\code{M90}}{a numeric vector}
    \item{\code{M91}}{a numeric vector}
    \item{\code{M92}}{a numeric vector}
    \item{\code{M93}}{a numeric vector}
    \item{\code{M94}}{a numeric vector}
    \item{\code{M95}}{a numeric vector}
    \item{\code{M96}}{a numeric vector}
    \item{\code{M97}}{a numeric vector}
    \item{\code{M98}}{a numeric vector}
    \item{\code{M99}}{a numeric vector}
    \item{\code{M100}}{a numeric vector}
    \item{\code{M101}}{a numeric vector}
    \item{\code{M102}}{a numeric vector}
    \item{\code{M103}}{a numeric vector}
    \item{\code{M104}}{a numeric vector}
    \item{\code{M105}}{a numeric vector}
    \item{\code{M106}}{a numeric vector}
    \item{\code{M107}}{a numeric vector}
    \item{\code{M108}}{a numeric vector}
    \item{\code{M109}}{a numeric vector}
    \item{\code{M110}}{a numeric vector}
    \item{\code{M111}}{a numeric vector}
    \item{\code{M112}}{a numeric vector}
    \item{\code{M113}}{a numeric vector}
    \item{\code{M114}}{a numeric vector}
    \item{\code{M115}}{a numeric vector}
    \item{\code{M116}}{a numeric vector}
    \item{\code{M117}}{a numeric vector}
    \item{\code{M118}}{a numeric vector}
    \item{\code{M119}}{a numeric vector}
    \item{\code{M120}}{a numeric vector}
    \item{\code{M121}}{a numeric vector}
    \item{\code{M122}}{a numeric vector}
    \item{\code{M123}}{a numeric vector}
    \item{\code{M124}}{a numeric vector}
    \item{\code{M125}}{a numeric vector}
  }
}
\examples{
data(SI_markers)
## maybe str(SI_markers) ; plot(SI_markers) ...
}
\keyword{datasets}
