\name{SI_traits}
\alias{SI_traits}
\docType{data}
\title{SI_traits}
\description{
data on traits for Selection Index Analysis
}
\usage{data(SI_traits)}
\format{
  A data frame with 420 observations on the following 18 variables.
  \describe{
    \item{\code{REP}}{a numeric vector}
    \item{\code{Block}}{a numeric vector}
    \item{\code{ENTRY}}{a numeric vector}
    \item{\code{MFL1}}{a numeric vector}
    \item{\code{FFL1}}{a numeric vector}
    \item{\code{EHT1}}{a numeric vector}
    \item{\code{PHT1}}{a numeric vector}
    \item{\code{GY1}}{a numeric vector}
    \item{\code{MFL.2}}{a numeric vector}
    \item{\code{FFL2}}{a numeric vector}
    \item{\code{EHT2}}{a numeric vector}
    \item{\code{PHT2}}{a numeric vector}
    \item{\code{GY2}}{a numeric vector}
    \item{\code{MFL3}}{a numeric vector}
    \item{\code{FFL3}}{a numeric vector}
    \item{\code{EHT3}}{a numeric vector}
    \item{\code{PHT3}}{a numeric vector}
    \item{\code{GY3}}{a numeric vector}
  }
}
\examples{
data(SI_traits)
## maybe str(SI_traits) ; plot(SI_traits) ...
}
\keyword{datasets}
