\name{ssa.resid}
\alias{ssa.resid}
\alias{ssa.resid.default}
\title{ssa.resid}
\description{
function for consolidating residuals from single site analysis
}
\usage{
ssa.resid(data, ssaResult, respvar, env, is.genoRandom = FALSE)
}
\arguments{
  \item{data}{a}
  \item{ssaResult}{a}
  \item{respvar}{a}
  \item{env}{a}
  \item{is.genoRandom}{a}
}
\author{Nellwyn L. Sales}
\keyword{residuals}