\name{stability.analysis}
\alias{stability.analysis}
\alias{stability.analysis.default}
\title{stability.analysis}
\description{
performs stability analysis
}
\usage{
stability.analysis(data, respvar, geno, env, method = c("regression", "shukla"))
}
\arguments{
  \item{data}{a}
  \item{respvar}{a}
  \item{geno}{a}
  \item{env}{a}
  \item{method}{a}
}
\author{
Alaine A. Gulles
}
\keyword{stability}
