\name{testQTLsegregation}
\alias{testQTLsegregation}
\alias{testQTLsegregation.default}
\title{testQTLsegregation}
\description{testQTLsegregation}
\usage{
testQTLsegregation(crossData, pvalCutOff)
}
\arguments{
  \item{crossData}{a}
  \item{pvalCutOff}{a}
}
\author{Rose Imee Zhella Morantte}
\keyword{analysis}