% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{themebasecolours}
\alias{themebasecolours}
\title{Colour set for theme - default ugly colours (6 base colours only)}
\format{Matrix with 4 columns and 6 rows - 4 tints (25\% steps) of 6 colours}
\usage{
data(themecolours)
}
\description{
Colour set for theme - default ugly colours (6 base colours only)
}
\keyword{datasets}

