% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{themecolours}
\alias{themecolours}
\title{Colour set for theme - 4 tints of 6 colours}
\format{Matrix with 4 columns and 6 rows - 4 tints (25\% steps) of 6 colours}
\usage{
data(themecolours)
}
\description{
Colour set for theme - 4 tints of 6 colours
}
\keyword{datasets}

