% Generated by roxygen2 (4.0.1): do not edit by hand
\name{timestampedfilename}
\alias{timestampedfilename}
\title{File name with timestamp}
\usage{
timestampedfilename(filename = "file", extension = "", folder = "",
  format = "\%Y\%m\%d_\%H\%M\%S")
}
\arguments{
\item{filename.}{String. Defaults to 'file'.}

\item{extension}{String. Defaults 'txt'.}

\item{folder}{Folder. String. Defaults to ''.}

\item{format}{Date-time format. String. Defaults to '%Y%m%d_%H%M%S'}
}
\value{
String
}
\description{
Return string file name including current time.
}
\examples{
timestampedfilename()
}
\keyword{date,}
\keyword{file,}
\keyword{save}
\keyword{time,}

