% Generated by roxygen2 (4.0.1): do not edit by hand
\name{timestampedfolder}
\alias{timestampedfolder}
\title{Folder name with timestamp}
\usage{
timestampedfolder(folder = "folder", path = "./",
  format = "\%Y\%m\%d_\%H\%M\%S")
}
\arguments{
\item{folder}{Folder. String. Defaults to 'folder'.}

\item{format}{Date-time format. String. Defaults to '%Y%m%d_%H%M%S'}
}
\value{
String
}
\description{
Create folder with name including current time and return the folder path.
}
\examples{
timestampedfolder()
}
\keyword{date,}
\keyword{file,}
\keyword{save}
\keyword{time,}

