\name{TTC_ME}
\alias{TTC_ME}
\docType{data}
\title{TTC_ME}
\description{
data for Triple Test Cross Analysis
}
\usage{data(TTC_ME)}
\format{
  A data frame with 180 observations on the following 5 variables.
  \describe{
    \item{\code{Env}}{a factor with levels \code{A} \code{B}}
    \item{\code{Block}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{Female}}{a factor with levels \code{F1} \code{P1} \code{P2}}
    \item{\code{Male}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9} \code{10}}
    \item{\code{Y}}{a numeric vector}
  }
}
\examples{
data(TTC_ME)
## maybe str(TTC_ME) ; plot(TTC_ME) ...
}
\keyword{datasets}
