% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{ifgbasecolours}
\alias{ifgbasecolours}
\title{Colour set for theme - IfG corporate colours}
\format{Matrix with 4 columns and 7 rows - 4 tints (25\% steps) of 7 colours}
\usage{
data(themecolours)
}
\description{
Colour set for theme - IfG corporate colours
}
\keyword{datasets}

