% Generated by roxygen2 (4.0.1): do not edit by hand
\name{LoadCustomThemes}
\alias{LoadCustomThemes}
\title{Load custom theme}
\usage{
LoadCustomThemes(mycols = themecolours, fontfamily = "Helvetica")
}
\arguments{
\item{mycols}{A vector of colours in hex (HTML) format. Defaults to ugly.}

\item{fontfamily}{Font family. Defaults to 'Helvetica'}
}
\value{
NULL
LoadCustomThemes()
}
\description{
Create custom ggplot theme based on a set of colours and a font family
}
\keyword{ggplot,}
\keyword{themes}

