\name{PBTools-package}
\alias{PBTools-package}
\alias{PBTools}
\docType{package}
\title{Plant Breeding Tools}
\description{An analytical software the performs statistical analyses that are used in Plant Breeding}
\details{
\tabular{ll}{
Package: \tab PBTools\cr
Type: \tab Package\cr
Version: \tab 1.5\cr
Date: \tab 2015-03-10\cr
Depends: \tab R (>= 3.0.2), lme4, nlme, multcomp, stringr, doBy, splines, MASS, Matrix, lattice, leaps, plotrix, DiGGer, qtl, qtlbim, ape, gdata\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Rose Imee Zhella A. Morantte <r.morantte@irri.org> and 
Nellwyn L. Sales <n.sales@irri.org>, with contributions from
Ms. Violeta A. Bartolome, Ms. Leilani A. Nora, Ms. Alaine A. Gulles
and Dr. Guoyou Ye of International Rice Research Institute

Maintainer: Rose Imee Zhella A. Morantte <r.morantte@irri.org> and Alaine A. Gulles <a.gulles@irri.org>
}
\keyword{ package }
