\name{QTL.analysis}
\alias{QTL.analysis}
\alias{QTL.analysis.default}
\title{QTL.analysis}
\description{QTL.analysis}
\usage{
QTL.analysis(crossobj = cross.data, QTL.path, geno, env.label = NULL, env = NULL, trait = "pred", step, method, threshold, distance, cofactors, window.size)
}
\arguments{
  \item{crossobj}{a}
  \item{QTL.path}{a}
  \item{geno}{a}
  \item{env.label}{a}
  \item{env}{a}
  \item{trait}{a}
  \item{step}{a}
  \item{method}{a}
  \item{threshold}{a}
  \item{distance}{a}
  \item{cofactors}{a}
  \item{window.size}{a}
}
\author{Rose Imee Zhella Morantte}
\keyword{analysis}
