% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SavePlot}
\alias{SavePlot}
\title{Load custom theme}
\usage{
SavePlot(plotname = "Plot", plotformat = "eps", ffamily = "Helvetica",
  splot = last_plot(), ploth = 21/2, plotw = 14)
}
\arguments{
\item{splot}{Ggplot object. Defaults to last plot}

\item{plotname}{String. Defaults 'Plot'.}

\item{ffamily}{Font family. Defaults to 'Helvetica'}

\item{plotformat}{String}

\item{plotw}{Width of plot in centimeters. Defaults to 14}

\item{ploth}{Height of plot in centimeters. Defaults to 11.5}
}
\description{
Save ggplot - ggsave wrapper. Saves image, GGplot object, and data.
}
\examples{
SavePlot()
}
\keyword{ggplot,}
\keyword{themes}

