% Generated by roxygen2 (4.0.1): do not edit by hand
\name{saveplot}
\alias{saveplot}
\title{Save plot with ggsave}
\usage{
saveplot(plotname = timestampedfilename(), plotformat = "eps",
  ffamily = "Helvetica", splot = last_plot(), ploth = 21/2, plotw = 14,
  plotdir = ".")
}
\arguments{
\item{splot}{Ggplot object. Defaults to last plot}

\item{plotname}{String. Defaults 'Plot'.}

\item{ffamily}{Font family. Defaults to 'Helvetica'}

\item{plotformat}{String}

\item{plotw}{Width of plot in centimeters. Defaults to 14}

\item{ploth}{Height of plot in centimeters. Defaults to 11.5}

\item{plotdir}{Directory for saving - plot, data and image are saved into subdirectories}
}
\description{
Save ggplot - ggsave wrapper. Saves image, GGplot object, and data.
}
\examples{
saveplot()
}
\keyword{ggplot,}
\keyword{themes}

