% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{themecolours}
\alias{themecolours}
\title{Colour set for theme}
\format{Matrix with 4 columns and 6 rows - 4 tints (25% steps) of 6 colours}
\usage{
data(themecolours)
}
\description{
Colour set for theme
}
\keyword{datasets}

