% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TintShade}
\alias{TintShade}
\title{Produce tints and shades for a set of colours}
\usage{
TintShade(colors, tints = c(), shades = c(), hexin = TRUE)
}
\arguments{
\item{colors}{Character vector with hex or RGB spec of each colour}

\item{tints}{Numeric vector specifying level of each tint, 0-1; 1 will be included automatically}

\item{shades}{Numeric vector specifying level of each shade, 0-1; 1 will be included automatically}

\item{hexin}{Are colours in hex format? Defaults to TRUE.}
}
\value{
Matrix; nrows=number of colours, ncol=number of tints and shades
includes the input colours
}
\description{
This function allows you to express your love of cats.
}
\examples{
TintShare(ifgbasecolours, c(0.75,0.5,0.25))
}
\keyword{colours,}
\keyword{themes}

